Refactor BloomFilterEngine for improved hash handling

+++ android/app/src/main/java/com/signalgate/multipoint/data/security/BloomFilterEngine.kt
@@ -1,39 +1,29 @@
 package com.signalgate.multipoint.data.security
 
-import java.nio.ByteBuffer
-import java.nio.ByteOrder
 import java.util.BitSet
 
 class BloomFilterEngine(
     private val expectedElements: Int = 500000,
     private val falsePositiveRate: Double = 0.02
 ) {
-    // Dynamically scale bit array size and hash functions based on mathematical optimization
     private val numBits: Int
     private val numHashFunctions: Int
     private val bitSet: BitSet
-
     private val lock = Any()
 
     init {
-        // Optimization equations: m = - (n * ln(p)) / (ln(2)^2)
         val m = (-expectedElements * Math.log(falsePositiveRate) / Math.pow(Math.log(2.0), 2.0)).toInt()
         numBits = if (m <= 0) 1000 else m
-        
-        // k = (m / n) * ln(2)
         val k = Math.max(1, Math.round((numBits.toDouble() / expectedElements) * Math.log(2.0)).toInt())
         numHashFunctions = k
-        
         bitSet = BitSet(numBits)
     }
 
-    /**
-     * MurmurHash3 implementation used for fast, uniform key distribution over 32 bits
-     */
     private fun murmurHash3(data: ByteArray, seed: Int): Int {
         var h1 = seed
         val length = data.size
         val nblocks = length / 4
+
         for (i in 0 until nblocks) {
             val index = i * 4
             var k1 = (data[index].toInt() and 0xFF) or
@@ -74,40 +64,32 @@ class BloomFilterEngine(
         return h1
     }
 
-    /**
-     * Maps an incoming number into the bit array using cascading seeds
-     */
     fun insert(phoneNumber: String) {
         val bytes = phoneNumber.toByteArray(Charsets.UTF_8)
         synchronized(lock) {
             for (i in 0 until numHashFunctions) {
                 val hash = murmurHash3(bytes, i)
-                val index = Math.abs(hash % numBits)
+                // Fix: Math.abs(hash % numBits) returns Long on some JVMs — cast to Int for BitSet
+                val index = (Math.abs(hash.toLong() % numBits)).toInt()
                 bitSet.set(index)
             }
         }
     }
 
-    /**
-     * Core Check Pipeline: If any bit evaluates to false, it is 100% definitively NOT in the list
-     */
     fun mightContain(phoneNumber: String): Boolean {
         val bytes = phoneNumber.toByteArray(Charsets.UTF_8)
         synchronized(lock) {
             for (i in 0 until numHashFunctions) {
                 val hash = murmurHash3(bytes, i)
-                val index = Math.abs(hash % numBits)
+                val index = (Math.abs(hash.toLong() % numBits)).toInt()
                 if (!bitSet.get(index)) {
-                    return false // 100% Certain False Positive Protection
+                    return false
                 }
             }
         }
-        return true // High probability match, verify via SQLite
+        return true
     }
 
-    /**
-     * Clear the filter cleanly when executing an atomic database swap
-     */
     fun clear() {
         synchronized(lock) {
             bitSet.clear()
